requirements.txt
torch
torchvision
tensorboard
pyyaml
tqdm
